--- fourtoon-cookie/src/components/character/CharacterList/CharacterList.styled.ts
@@ -1,20 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-    listContainer: {
-        flex: 1,
-        flexGrow: 1,
-        flexDirection: 'row',
-        alignItems: 'center',
-        alignContent: 'flex-start',
-        padding: 0,
-        width: "100%",
-        height: "auto",
-    },
-    columnWrapper: {
-        flexDirection: 'row',
-        justifyContent: 'flex-start',
-        width: "100%",
-        height: "auto",
-    },
-});
